+++ commonroad_crime/measure/time/ttm.py
@@ -57,10 +57,14 @@ class TTM(CriMeBase):
         self._maneuver = maneuver
 
     def visualize(self, figsize: tuple = (25, 15)):
-        self._initialize_vis(figsize=figsize, plot_limit=utils_vis.plot_limits_from_state_list(self.time_step,
-                                                                                               self.ego_vehicle.prediction.
-                                                                                               trajectory.state_list,
-                                                                                               margin=10))
+        if self.configuration.debug.plot_limits:
+            plot_limits = self.configuration.debug.plot_limits
+        else:
+            plot_limits = utils_vis.plot_limits_from_state_list(self.time_step,
+                                                                self.ego_vehicle.prediction.
+                                                                trajectory.state_list,
+                                                                margin=10)
+        self._initialize_vis(figsize=figsize, plot_limit=plot_limits)
         self.ttc_object.draw_collision_checker(self.rnd)
         for veh in self.sce.obstacles:
             if veh is not self.ego_vehicle:
@@ -78,7 +82,7 @@ class TTM(CriMeBase):
         if self.value not in [math.inf, -math.inf] and self.ttc:
             tstm = int(utils_gen.int_round(self.value / self.dt, 0)) + self.time_step
             utils_vis.draw_state(self.rnd, self.ego_vehicle.state_at_time(tstm), TUMcolor.TUMgreen)
-            tstc = int(utils_gen.int_round(self.ttc_object.value / self.dt, 0))
+            tstc = int(utils_gen.int_round(self.ttc_object.value / self.dt, 0)) + self.time_step
             utils_vis.draw_state(self.rnd, self.ego_vehicle.state_at_time(tstc), TUMcolor.TUMred)
 
             tstc = int(utils_gen.int_round(self.ttc / self.dt, 0)) + self.time_step
@@ -122,17 +126,19 @@ class TTM(CriMeBase):
         """
         ttm = - math.inf
         low = initial_step
-        high = int(utils_gen.int_round(self.ttc / self.dt,  str(self.dt)[::-1].find('.')))
+        high = int(utils_gen.int_round(self.ttc / self.dt + self.time_step,  str(self.dt)[::-1].find('.')))
         while low < high:
             mid = int((low + high) / 2)
             state_list = self.simulator.simulate_state_list(mid)
-            utils_gen.check_elements_state_list(state_list, self.dt)
-            self.state_list_set.append(state_list[mid:])
             # flag for successful simulation, 0: False, 1: True
             flag_succ = state_list[-1].time_step == self.ego_vehicle.prediction.final_time_step
+            if not flag_succ:
+                high = mid
+                continue
+            self.state_list_set.append(state_list[mid:])
             # flag for collision, 0: False, 1: True
             flag_coll = self.ttc_object.detect_collision(state_list)
-            if not flag_coll and flag_succ:
+            if not flag_coll:
                 low = mid + 1
                 self.selected_state_list = state_list
             else: