+++ tutorials/evaluate_esmini.py
@@ -4,7 +4,8 @@ import commonroad_crime.utility.visualization as utils_vis
 
 def main():
     scenario_id = "OSC_CutIn-1_2_T-1"
-
+    # time_steps=[20, 34, 99])
+    scenario_id = "OSC_Overtake-1_1_T-1"
     # ==== build configuration
     config = ConfigurationBuilder.build_configuration(scenario_id)
     config.update()
@@ -13,7 +14,7 @@ def main():
     # ==== visualize the scenario at given time steps
     utils_vis.visualize_scenario_at_time_steps(config.scenario,
                                                plot_limit=config.debug.plot_limits,
-                                               time_steps=[20, 34, 99])
+                                               time_steps=[69, 121, 129])
 
 
 if __name__ == "__main__":